# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Candela RMS API
  version: 1.0.0
  description: Retail Management System API

servers:
  - url: http://localhost:5000/api
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, MANAGER, CASHIER]
        storeId:
          type: string
        createdAt:
          type: string
          format: date-time

    Product:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        barcode:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        unitPrice:
          type: number
        costPrice:
          type: number
        taxRate:
          type: number
        minStock:
          type: integer
        storeId:
          type: string

    Sale:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        subtotal:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

paths:
  /auth/login:
    post:
      summary: User login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string

  /products:
    get:
      summary: Get all products
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: storeId
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      summary: Create product
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        201:
          description: Product created

  /pos/sale:
    post:
      summary: Create new sale
      tags: [POS]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                paymentMethod:
                  type: string
                customerId:
                  type: string
                discount:
                  type: number
      responses:
        201:
          description: Sale created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'

  /reports/sales:
    get:
      summary: Get sales report
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: storeId
          in: query
          schema:
            type: string
      responses:
        200:
          description: Sales report data